Move engine util.js to src
